% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NERD_function.R
\name{NERD_function}
\alias{NERD_function}
\title{The Nosocomial Evolution of Resistance Detector (NERD)}
\usage{
NERD_function(times, res_info, param = 365)
}
\arguments{
\item{times}{The date information. Should be stored in non-negative integers (e.g. days).}

\item{res_info}{The resistance information. Should be stored on a discrete scale from 1 to some value m. Values below 0 are discarded. For us: load "CData" which stores evolution information and select appropriate column to extract antibiotic of interest.}

\item{param}{The parameter for the Poisson part of the test. Defaults to 365 days = 1 year.}
}
\description{
The NERD package analyses temporal trends in antibiotic resistance data.
}
\examples{
NERD_function()
}
\keyword{Antibiotic}
\keyword{Resistance}

